Handle missing local_date fields while expanding the data_df
While we expect the local_date fields to be filled out in general, it is an
expected use case for places to not have the end_local_dt set for arbitrarily
long time, until the next trip starts. We should not expect to use those
fields, but we should expect to parse the existing fields correctly.

This is a simple change that allows us to do that.