Simplify keys used to choose metrics
The keys for the time series follow the general key pattern of category/name
and are `metrics/daily_user_count`, etc. But the keys for the API call don't
need to be the same as the metadata keys, and in fact, should not be the same,
since the call through the API can be at a user specified frequency.

Let's switch to simpler names for the API instead - `count`, `duration`,
`distance`, etc