Compute the section information as part of the test
It turns out the prior iterations of the test were passing because of old data
in my database. The tests failed in jenkins because no sections were found.
https://amplab.cs.berkeley.edu/jenkins/job/e-mission-server-prb/591/

On re-running after purging data, the tests failed in the same way.
So I added in code to compute the sections as part of the test.
Since it might be useful to do this elsewhere, put the code into common.