Prevent extra_queries from overriding regular queries
While investigating
2f7fbe1df73f0611b97cc2e71fbc89d57216727c
I wasn't quite sure how the tests had passed before. After all, we were only
querying for sections with the UUID from the test - how would the presence of
old data in the database affect it.

The answer was both subtle and annoying. It turned out that since we used both
the user_id for the user and the exclusion query for the test phones in the
SAME query, the exclusion query overwrote the query for the user id. So
although we passed in a user_id to the query function, we were eventually
running an aggregate query that returned everything except the test phones!!

In order to fix this:
- I realized that we only cared about excluding the test phones during the
  aggregate query, since for non-aggregate timeseries, we filtered for a
  particular phone anyway. And we had defined the exclusion function in the
  aggregate timeseries anyway. So why not make that the user_query for the
  aggregate timeseries by default!?
  Much cleaner code and can effectively turn into a private function
- I also checked for overlap of top level keys between existing queries and
  extra queries and raised an exception if they did overlap. I also added a
  test case to check that this behavior works as expected. Hopefully, this will
  ensure that we don't have such subtle bugs in the future.