Handle the case where no sections are matched in the query gracefully
In 2f7fbe1df73f0611b97cc2e71fbc89d57216727c, we computed sections as part of
the test so that we could avoid issues caused by having no sections. But having
no sections match the query is a perfectly valid use case - user enters 2000 -
2001, I don't go to Berkeley for a week, etc.

So we should deal with the case robustly and not crash. Add a simple check for
the section_df size to avoid unforseen crashes, and add a new test that queries
from 2000 -> 2001.