Add support for returning the number of users contributing to each metric
As we flesh out the API for the metrics, we return user and aggregate metrics.
But for the aggregate metrics, it would be useful to return not just the
metrics, but also the number of users that it is based on so that:
- we can see how valid it is
- we can compute the mean of the values for comparison with the individual
  values

This commit adds support for returning the number of users contributing to any
metric. In particular:
- we add a new field to store the value
- we include user_id in the data_df and change the test case for it to expect
  21 columns instead of 20
- we use `unique()` on the new `user_id` column to compute the number of users
  and store it in the field
- we modify the tests to ensure that the value is returned, and that it makes
  sense