Better tests continue to uncover subtle bugs in the extra_query code
While running automated tests with the extra_query check, we got an error
claiming that the data.mode query was being overwritten.

```
======================================================================
ERROR: testCommutePopRoute (netTests.TestVisualize.TestVisualize)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "<https://amplab.cs.berkeley.edu/jenkins/job/e-mission-server-prb/ws/emission/tests/netTests/TestVisualize.py",> line 48, in testCommutePopRoute
    self.day_end_dt, None)
  File "emission/net/api/visualize.py", line 79, in range_mode_heatmap
    extra_query_list=extra_query_list)
  File "emission/storage/decorations/analysis_timeseries_queries.py", line 44, in get_entries
    doc_cursor = ts.find_entries([key], time_query, geo_query, extra_query_list)
  File "emission/storage/timeseries/builtin_timeseries.py", line 158, in find_entries
    sort_key)
  File "emission/storage/timeseries/builtin_timeseries.py", line 168, in _get_entries_for_timeseries
    extra_query_list))
  File "emission/storage/timeseries/builtin_timeseries.py", line 90, in _get_query
    list(overlap_keys))
AttributeError: extra query would overwrite keys ['data.mode']

----------------------------------------------------------------------
```

This was bizarre because we only passed the data.mode field in once during the
extra_query.

On digging deeper it turned out that the implementation of _get_query()
modified the user_query so that subsequent calls to _get_query() would not
start empty but build on the previous query!

Fixed that and the test passes now.