Hook up the metrics to the UI
Woot! Woot! We finally hook up the metrics to the UI.
The change to the API layer is fairly simple - we just call the methods in the
metrics wrapper.
The change to the webapp is also simple, however we do a bit of restructuring
at the same time. In particular
- we move the heatmap code out to its own javscript file
- copy the heatmap javascript and HTML as the basis for the metrics
- modify them to match the metrics
    - remove the hour and day filters, which are confusing
    - instead of mapping the returned lat,lng points, map the stats objects to
      the format expected by nvd3

Existing bugs:
- y axis label does not get updated when metric changes. This can probably be fixed
    by setting the config `refreshDataOnly` to false
    https://krispo.github.io/angular-nvd3/#/quickstart
    or by forcing a refresh `scope.api.update();`
    https://github.com/krispo/angular-nvd3/issues/459
- should scale values properly so that they are understandable (m -> km, m/s ->
  km/hr), etc

Once the tests pass, I'm going to merge this