Get a reference to the database once outside the loop and reuse it
In several places in the code, while we retrieved entries from the database in
a loop, we would call `get_section_db()` or `get_trip_db()` within the loop.
While this is less of a concern now that we make one connection at the time of
module load, it is still a good cleanup.

Note that some of these also have the logging change included, but a majority
of the logging changes are in the next commit.