Move all logging to the main function
Before this change, we logged the message, but did not log any timestamps,
which makes it harder to see how long various operations took, correlate with
logs on other servers, etc.

The logs were also configured individually in each test and some non-test
modules, making it hard to change the log format easily. Finally, the logs
were occasionally configured globally on module load. This meant that they
would override any configurations in main.

I went through and unified the logging module configuration.
- modules do not have module level logging configurations. no exceptions.
    (scripts in `bin` are exceptions because they are intended to be more like scripts)
- modules that are invoked directly by the user (e.g intake, webapp) have
  existing configurations
- tests have a standard logging configuration defined in `emission.tests.common`
- all tests configure their logging in main using the standard logging configuration

The exception is cfc_webapp. As part of this change, I created a __main__
function, and the app starts up correctly with it, but if I move the config in
there, it is never applied. In particular, the `This should go to the log file`
never shows up.

Some other module must be configuring logging, but I am not sure which it is,
since a search for logging.basicConfig() in the emission directory does not
find any matches that are not commented out or in `prior_unused` or `notebooks`
or `incomplete_tests`.