Fix regression in section segmentation on iOS caused by expansion of local_dt
In de06821e5c2a2a87bd4ac246062720d698fc5d04, we expanded the local_dt into its
constituent elements while creating the dataframe from a set of objects. We
converted all existing tests to match the specification. However, it turns out
that the segmentation tests only exercise the android code path, which is
different from the iOS code path.

In this change, we enhance the segmentation tests to work on iOS as well. As
expected, the tests failed with the same error as the server.

```
Traceback (most recent call last):
  File "/Users/shankari/e-mission/e-mission-server/emission/analysis/intake/seg
mentation/section_segmentation.py", line 42, in segment_current_sections
    segment_trip_into_sections(user_id, trip.get_id(), trip.data.source)
  File "/Users/shankari/e-mission/e-mission-server/emission/analysis/intake/seg
mentation/section_segmentation.py", line 73, in segment_trip_into_sections
    segmentation_points = shcmsm.segment_into_sections(ts, time_query)
  File "/Users/shankari/e-mission/e-mission-server/emission/analysis/intake/segmentation/section_segmentation_methods/smoothed_high_confidence_with_visit_transitions.py", line 87, in segment_into_sections
    location_points.iloc[0]), motion_changes[0][1])
  File "/Users/shankari/e-mission/e-mission-server/emission/analysis/intake/segmentation/section_segmentation_methods/smoothed_high_confidence_with_visit_transitions.py", line 44, in extend_activity_to_location
    'local_dt': location_point.local_dt,
  File "/Users/shankari/OSS/anaconda/lib/python2.7/site-packages/pandas/core/generic.py", line 2246, in __getattr__
    (type(self).__name__, name))
AttributeError: 'Series' object has no attribute 'local_dt'
```

The error is caused because we are passing in the dataframe row directly, and
etrying to set the `local_dt` for an entry from it. The solution is similar to
bce9c3425e3806e07aba8d1b6be0ff03aae5afb0, we retrieve the entry corresponding
to a row and return it.