Return the correct field for querying smoothing entries
The entries are smoothed by querying for sections and then smoothing the
sections. As part of 96ed8711135f7d4cae0706d5079c8dbff8a02b86, we converted the
queries to work on any field instead of fields within data, which meant that
the fields had to be specified as `"data.ts"` instead of `"ts"`. We changed the
field for `SECTION_SEGMENTATION` to `data.end_ts`, but didn't change the query
for the smoothing. This meant that we have not been running smoothing properly
since Apr.

This resolves
https://github.com/e-mission/e-mission-server/issues/334