Create the timeline with a correctly sorted list of trips
This fixes https://github.com/e-mission/e-mission-server/issues/345

The long and incredibly detailed explanation is in there. Basically, if we use
the values from a map directly, they are not sorted, and so the first trip is
not first in the list, we think that the timeline begins with a place instead
of a trip and follow the timeline, which means that the first trip is dropped.