We have enough points to make a decision only when they span at least 5 mins
Before this change, we had two checks to determine whether we had enough
points to make a decision

- whether we had at least 10 points
- whether any of them were in the past 5 mins

If the points that we received were 30 secs apart, as we expected, this would
work just fine. But if they were more frequent, as when navigating while
walking, then we detect a trip end way before 5 mins.

In this change, we also add an additional check to ensure that the points from
the last 5 mins query actually span 5 mins. With this check, I think that we
can remove the last 10 points check, but I want to wait until we have a more
principled checker to confirm before making a subtractive change.