Don't check for continuation when we are starting a new trip
This fixes #302. Sometimes, we get some points beyond the expected end of a
trip. We want to ignore them so we treat them as a continuation of the current
trip. However, if the points occur at the beginning of a trip, we don't want
this to happen.

Note that we have `justEnded = true` at the beginning of a trip, to have a
ingle control flow for the start and restart cases. So we need to have an
additional check for the idx before we detect a continuation.