Handle mismatch between phone and server trip end detection
This is a partial fix for #352. If there is a mismatch between phone and
server trip end detection, then we can end up with "joined" trips. In order to
work around that, we detect if there is a large gap between two points and if
there is, we detect that a visit has started.

More discussion is in the bug, but we also check the distance covered in that
gap (to ensure that we don't truncate trips with poor quality) and ensure that
when we detect a trip end in this new way, we also start a new trip at the
same time.