Handle the (corner) case in which there is only one high confidence valid motion activity
This is a partial fix for #356. Basically, if we have only one high confidence valid motion activity, we still want to treat that as a section, since it will ensure that the trip is not squished later.

A more complex fix to avoid squishing real trips even if they have no high
confidence motion activies has been shelfed because of complexities in
implementation and concern that it could cause spurious trips to show up
again.