Dynamic determination of what constitutes "the same spot"
This is a follow up fix for #352. Recall that our fix is to split trips if
there is a large gap in time and a small gap in distance.  In
69aa9f0a5cd1e0b7ee420632741ad2411b5244d1, our definition of "a small gap in
distance" was 2 * the distance filter (200m). Since the geofence radius is
100m, we assumed that most real start points will fit this criterion.

However, it turns out that the trip to pick up the kids on the 21st does not -
the distance was over 200m.

```
2016-08-17 21:39:38,177:DEBUG:last10PointsDistances = [ 229.10965712  223.17078778  229.7894694   227.49020288  225.23634705
  224.49880325  226.45225137  217.05680638  217.05680638    0.        ] with length 10, shape (10,)
```

Therefore, we change the check to be for speed instead. If you cover less than
100m/5 mins, then you are basically stopped. In this case, we covered 200 m in
way more than 10 mins (in 75 mins) to be precise, so we do mark it as a new
trip.

This also contains a fix for duplicate entries (timeDelta = 0) which has
issues while computing the speed.