Add a new test suite that uses data from known days to verify the end to end intake run
Test is fairly self-explanatory.
It has the raw input timeseries for a set of days that exercises several of the segmentation issues and tweaks that we have put into place.
It has the expected output timeline.

It compares them.

The goal is to be able to more quickly tweak the segmentation code without
worrying whether it is breaking something serious (like stupid #345). So it is
not really a unit test, but more of an end to end regression test.