Refactor timeseries to allow creation of a dataframe from an arbitrary list and dedup it
Before this, the timeseries had a method to retrieve the entries for a query
and the dataframe for a query. However, sometimes, we want to retrieve the
entries, manipulate them and then convert them to a dataframe. This commit
refactors the code to allow an arbitrary iterator of entries to be converted
to a dataframe.

It also dedups the entries in the dataframe. This is done by getting the list of properties for the objects in the dataframe and calling `drop_duplicates` on it. This works around the issues with duplicate entries that we saw in practice
(https://github.com/e-mission/e-mission-server/issues/288#issuecomment-242540929, https://github.com/e-mission/e-mission-server/issues/288#issuecomment-242615743)

Using the new method, as opposed to arbitrary list -> df conversions ensures
that the dedup is called correctly.

We also change all existing instances (except in the clean and resample code)
to use the newly created function.