Change the clean and resample code to work on deduped dataframes
In
https://github.com/e-mission/e-mission-server/issues/288#issuecomment-242539401,
we found that duplicate entries caused us to have serious issues during
cleaning and resegmentation. It is very possible that such changes are also
the reason why prior outlier detection, etc have not worked well.

We have found instances of this as late as Jun. We need to investigate why
this happens. But in the interim, we need to make our code more robust to it.
Therefore, we dedup entries before returning the corresponding dataframes
in the timeseries.

However, the clean and resample code wasn't getting the dataframe in the
standard way. Instead, it was using a mishmash of lists and dataframes, and
converting from lists to dataframes in a non-standard way.

This refactors the clean and resample code to start with dataframes and then
use them throughout.

It also changes the notebook to use the new, standard dataframe methods