Revert WORM -> RO change
In 60080f098bd96e6fea6ca2acad391548c4ff56c7, we changed the location wrappers
from read-only in order to support the creation of a dummy location to
represent the extrapolated first point of the trip. However, in
771189c13f4302895f8fe91496579268841d80c1, we refactored the cleaning code to
use dataframes throughout. So we don't really need this anymore.

Let's go back to the read-only entries and ensure that we don't inadvertently
overwrite location entries.