Fix over extrapolation by ensuring that there is a gap during the trip end
As seen in
https://github.com/e-mission/e-mission-server/issues/288#issuecomment-242531798,
if a trip was slow earlier and fast later and there was no tracking for the
slow part, we will end up with an extrapolated start time that is before the
end of the previous trip.

We check and fix this condition and add a new test case for it.