Adjust the timeline to return both trips and untracked time
Untracked time objects are used to fill gaps in the timeline where we have to
information. An untracked time period is between two places, just like a trip
- in effect, it is a "ghost" trip that we have no information on, but we can
deduce from seeing the effect (moved from place A to place B). This means that
the timeline decorator needs to handle untracked time so that we can iterate
corectly over it.

- Add support for passing in an untracked key similar to trip and place keys
- Change the queries to read the objects for the timeline to include the untracked key as well