Handle untracked time downstream
In general, we don't want untracked time to count towards any analysis - no
heatmap, no metrics. That's why we created them as a separate data type with a
separate stream. So there were few changes required to the code.

The one place that the untracked time does count is in the timeline. So we
build on 7ff18ea79a54d2b1be9bfa167bef1fab46e57407 to ensure that the untracked
time is included when we go from the raw timeline to the cleaned timeline, and
then when the cleaned timeline is converted to geojson to send to the server.