Fix the tests to support untracked time
We use the data from 27th Aug 2015 for a lot of tests, and the data for that
day includes a restart. Since we now handle restarts properly, we need to fix
the tests to support them as well. Fortunately, since untracked time is only
important for the timeline, the only tests affected are the segmentation
itself, the timeline code and its usages in export.