Add tests for air as well
Note that the current implementation loads the data as multiple users and runs
the pipeline separately - the result is slightly different if all data is
loaded for a single user then the first trip on the 4th gets smushed with the
last trip on the 27th. Will file a separate issue for that and fix it
separately.