Stop handling the case in which the last point in a section is not the section end
This was caused by an error in resetting the section end in the case where the
last few points of the sections are filtered. Now that we fixed it
(65859374c35d5794c00cdd109aca842f0f9acb54), we don't need to TODO any more.

We replace it with an assert, and assuming the assert does not trigger while
running the pipeline against my data, I can remove the if check too!