Add an UNKNOWN section if the distance from the previous place is greater than a threshold
In prior commits, we have attempted to fill the holes in timelines caused by
short trips without sufficient high confidence motion activity entries to
create a proper section (6aac0d8b6fe13ceb2acc09c6018ac2bd9d169957, 01043f188803053dd951d720c560b5143f52ce9a). However, at least on android, we didn't handle
the case in which there were zero high confidence valid actions (https://github.com/e-mission/e-mission-server/issues/356#issuecomment-240630757).

But while investigating further, we did find such a point
(https://github.com/e-mission/e-mission-server/issues/385#issuecomment-244826007).
So we change the code to compute the distance from the previous place, pass
that it, and compare it against a threshold to determine whether we should
create an UNKNOWN section anyway or not.

Confirmed using Gabe's data that this does not re-introduce fake/spurious
trips.

It is not clear that this is required on iOS, because we already check for
visit transitions. But it is not a bad check to add, and we can tweak it
further later after testing with @juemura's and @sunil07t's trips.