Fix subtle-but-annoying issue with squishing trips
When we squish zero section or < 100m trips, we effectively merge their start
and end places. But the start and end places don't have to be identical, they
can be 100m apart. This change fixes that discrepancy by moving the end place
(and in effect the start place of the next trip) slightly to match the start
place. It was a ton of work to handle what is a very small distance, but is
potentially visually noticeable on short trips.
https://github.com/e-mission/e-mission-server/issues/385#issuecomment-244793203

The steps to fix it are:
- calculate distance delta and new start_ts
- set the trip start information to the new location and time
- change the trip metrics using the deltas
- find the first section of the trip
- set the section start information to the new location and time
- change the section metrics using the deltas
- change the section speed and distance arrays by inserting values for the new segment
- add a new point for the new start location to the beginning
- modify the previous first point to have the correct distance and speed (previously 0 because it was first)