Better of the zero median speed case + fix regression in it + make asserts really work
if the median is zero, the time can be zero even if the distance is not zero.
Need to change the check to account for that.

Also try to handle the zero median case better by taking the median non-zero value.

Also fix the asserts (which didn't appear to catch the case in which I was using the start_ts after the override (89c71365e3d2594806d750259c8d023737f035d4)).