Fill in fake data on iOS in order to more easily detect jumps
This fixes #391

Otherwise, because of the distance filter, we may have too few points to
detect what is a jump and what isn't. This change makes the iOS data more
android-like by filling in noisy points during times when there was a big gap
in the data collection. We can fill them in with noise that falls within the
distance filter and make it be consistent with the observed values.