Squish untracked time sections if necessary
While fixing a56adddc5dc8c94cbe98964aafb17df3bc3f724c, we were able to filter
out the jump, but the trip still started from Locksunart Way. This is because
we detected untracked time during the first jump _to_ Locksunart Way and then
since everything else was a squished trip, kept mapping places to it although
they were actually fairly far apart.

More details are in
https://github.com/e-mission/e-mission-server/issues/391#issuecomment-247375367

The solution is to find the next un-squished trip for every untracked time,
and squish the untracked time if its start and the next unsquished trip start
are "close".

This works for the 20th.

This fixes #391