Add in a redirect URL
This is a workaround to sending custom URL schemes via gmail or social networks
such as facebook and twitter. All such custom URL schemes are stripped out by
the networks. In response, we use a redirect link on our server that redirects
to the custom URL with the same route and parameters.