Check to see that the requested UUIDs are actually public
Even for the current UUID-based getPublicData API call, we want to check that
the requested UUIDs are actually public. We don't guarantee that UUIDs are               private, for example, my UUID is strewn across the github repo in a bunch of             tests. We are planning to share UUIDs with the bic2cal researchers, for example.

So let's add a simple check to see if the UUIDs are known to be private before           returning them. This is required before I deploy                                         75f591f3ae1659cc1ce69cc23bbd888e264b7462to the real server, which I plan to do
in a little bit.