Make iOS tests robust again
This fixes #410. We set a random seed, so the randomness is predictable and we
get the same values. Might consider doing this even if it is not the tests for
the nice property of deterministic execution. Note that we don't actually need
true randomness, we just need some noise so that we don't get a lot of zero
values. pseudo random values are good enough.