Restore logging by default
For a very long time, we have not been able to get logging to show up by
default, particularly on the jenkins tests, and I have always wondered why.
Today, I spent some time poking around, and it looks like the problem was
`get_database.py` which included `archiver.py`, which threw an exception on
startup from `get_conf`.

I am a bit fuzzy on who exactly was calling `get_conf` - there was nothing in
the stacktrace, but we're going to restructure the communication with the
archiver to be... an archive anyway. Immediate stats will be in the timeseries
database just like always and long-term stats can migrate to the archiver.

So I remove the archiver code from `get_database.py`, change the stats code to
write directly to the local mongodb timeseries, and then logging works!