Only perform accuracy filtering for continuous collection
For geofencing, we filter the accuracy on the phone itself in order to
determine when the trip ends. But for continuous sensing, we have not done it
so far in order to accurately measure the power drain on the phone. So then we
have to filter on the server. This is a corner case because almost all real
users will use geofencing.

However, the filter_accuracy step, probably because of our algorithm, is the
single most expensive step in the pipeline.

https://github.com/e-mission/e-mission-server/issues/408#issuecomment-248052863

So we should only call filterAccuracy if we have continuous sensing turned on.
However, determining if we have continuous sensing turned on is complex
https://github.com/e-mission/e-mission-server/issues/412

So for now, we assume that all non-public phones, e.g. non-test phones are
geofenced and do not require accuracy filter checking.