Support bulk insert of entries and create a global ref to the collections
The next big component of the time was the time to open
connections to the database, and the re-indexing caused by the multiple
inserts.

https://github.com/e-mission/e-mission-server/issues/408#issuecomment-248066120
https://github.com/e-mission/e-mission-server/issues/408#issuecomment-248069629

We fix these by implementing a new bulk_insert operation that can either have
the underlying collection passed in (fast) or inferred (slower). We also move
the initialization of the connections to the one time load of the class.

After this, there does not appear to be any low-hanging fruit left.
https://github.com/e-mission/e-mission-server/issues/408#issuecomment-248146797