Switch the input usercache processing to use bulk_insert as well
We already saw from the profiling
https://github.com/e-mission/e-mission-server/issues/408#issuecomment-248065051
that lots of individual inserts are slow, and we fixed the
`clean_and_resample` step to use bulk insert instead.

We use bulk insert in the user cache move to long term as well, although that
does not show up in the profiling since we start with entries in long term
during testing.

Let's pre-emptively change the usercache to use bulk inserts as well