Drop na before determining the different filter types
We found this minor issue while running against real data.
If we have an invalid line in the loc_df (unsure why this would happen), we
end up with a filter list which looks like this ["time", nan] and fails down
the road.

https://github.com/e-mission/e-mission-server/issues/407#issuecomment-248222549

We simply drop the n/a before proceeding.