Allow loading the same data multiple times to generate multiple timelines
By, default, as highlighted in the README, loading the same data multiple
times leads to old data being overwritten. So if you want to load the same
data multiple times to fake having other users, you are out of luck.

Now that we create a new UUID and register the user as part of the load
anyway, we also add a -n option to create a new timeseries instead of
overwriting any previous one.

Also updated the README.