Parallelize execution to ensure that we can really handle 500 users
This finally fixes #408

The changes in https://github.com/e-mission/e-mission-server/issues/413 are
awesome and have reduced the runtime to 30% of the (admittedly unprofiled)
original. But note
(https://github.com/e-mission/e-mission-server/pull/413#issue-178121113) that
we can only scale to 300 users, not 500. In order to do that, we need to
parallelize - at least two processes, maybe three.

Fortunately, we have already structured the code as a nice pipeline and the
processing for one user is completely separate from the processing for other
users, so it is trivial to parallelize. We change the intake pipeline to be run
for a batch of users, add multiprocessing to run multiple processes in parallel
and write a naive scheduler that distributes the users among the processes
based on the number of messages to process for each them.

While running this, we discover that calling bulk_write with an empty array
throws an exception, so we add a new timeline method to check for empty and use
it to skip the write in that case.

We also had to add a new method to get the message count to the usercache
because the existing message method returns a list, and I don't want to
refactor right now (see comments for details).