Fix a bunch of minor issues in the cleaning
- Check for zero length in all bulk_insert calls
- Add sanity checks for erroneously classified walk and bike sections
    https://github.com/e-mission/e-mission-server/issues/407#issuecomment-248524098
- Handle the case in which all points are filtered by adding a single point
  corresponding to the start place enter.
    https://github.com/e-mission/e-mission-server/issues/407#issuecomment-248514593
    https://github.com/e-mission/e-mission-server/issues/407#issuecomment-248538469
- We are supposed to be squishing together places that are close to each
  other. For some reason, we squish together places that are far from each
  other. Why isn't the interim untracked time? Dunno, but add a sanity check to
  only squish if the places are near each other. We end up with a trip that has
  a big blank gap, but that is better than a 68km long walk trip
- Fix trip segmentation to detect a trip end if the gap is more than 12 hours
    https://github.com/e-mission/e-mission-server/issues/407#issuecomment-248635443

Also fix the iOS segmentation code to be similar to android in terms of
handling missing points. We don't want to end the trip every time there is a
gap if we have been moving in that time.

We need to re-run the whole pipeline after this because of the segmentation
changes.

Also, there is a minor fix to the tests because after the change, the zigzag
around the gym becomes one section.