Remove sources of log spew
Downgrade two of the three fairly innocuous errors in
https://github.com/e-mission/e-mission-server/issues/407#issuecomment-248974661

One is a normal use case and is marked as such. The other is downgraded to a
warning, and the handling of the DuplicateKeyError is removed because we don't
insert until outside the loop, where we use a bulk_insert.

The third will be moved with the upcoming `clean_and_resample` changes.