Fix the issue in which all section points are filtered
In 44f33af0012f6457064060b8f01acce07806329b, we fixed this for the first
section in a trip. This fixes it for all sections. Since the trip that caused
this in production runs fine locally, so the error is not reproducible, also
write a test case that forces use case to happen and tests that it works
correctly.

Other minor fixes:
As promised in 3f985700e5590e208db2b1e47daead1ca5c6b898, remove the third source of log spew by downgrading to info.

Update to match the new location of `compare_rounded_arrays`, changed in c506f14261409bf7cf998036872ceb3f25af1bdd