Bunch of fixes to iOS smoothing
Based on some changes that we made recently, the previous jump trip was merged.
This caused a regression in the tests and while investigating it, we found
several issues in the iOS fake point addition:

- we were only filling the first gap
- we were returning after the first big gap and not filling things later.

Even after fixing those, we needed to make other changes to support the iOS distance filter. Details are at:
https://github.com/e-mission/e-mission-server/issues/391#issuecomment-249112836

As a bonus - this also seems to have fixed some prior jumps - on the production server, we had this long bike trip

```
(UUID('ee6dadba-53b6-421f-94d1-27bc96e023cf'),
 u'2016-06-21T22:15:45.521593-07:00',
 147838.28720068303,
 15356.997407197952)
```

But on pulling and running it locally, I don't see that trip any more - last
trip ends at 10:07.