Remove the group user registration from the habitica tests
We used to register the group party leaders at the beginning of each test.
First, we do not need to do this any more since we are not autocreating parties.

Second, it won't work anyway because the group leaders (including me) will soon
have passwords that are not "autogenerate_me".