Don't store zombie recreated locations to the database
This fixes #421. Basically, we were always storing the recreated locations for
sections to the database, even if the trip was a spurious trip and we skipped
storing the trip and the associated sections. This led to zombie recreated
locations which were in the db but had no associated section.

Normally these would not be a problem, but in the aggregate heatmap, we retrieve all recreated locations, so they would throw off the heatmap. Let's just return both the section and the associated points and store both of them together.