Detect restarts in the middle of a trip as well
Refactor the restart logic into its own module and use it to detect restarts during a trip as well. If we have a large time gap, instead of looking at only the speed gap, which can generate spurious trips if the phone has been off for a sufficiently long time, we also reuse the restart checks to determine whether to split the trip or not.

See https://github.com/e-mission/e-mission-server/issues/407#issuecomment-249490852 and https://github.com/e-mission/e-mission-server/issues/407#issuecomment-249495997 for examples.

Also make minor changes to the ground truth to reflect this.