Implement framework for autochecking tasks
Note that the TestHabiticaAutocheck test has NOT been updated properly.
It will FAIL as is.

This sets a framework for defining a small language for autochecked tasks.

The framework will load a mapping function based on the language, and call it
with the user_id, the task (that contains task arguments) and the previous
state.

It returns the new state, which is then persisted back to the database.

Rework the existing walk and bike task into this framework
and change instances that called it to call the new code instead.