Handle the case in which there are no transitions
potentially for continuous sensing, we may have no transitions in the timeframe
that we are considering. So we can't show columns from it because the dataframe
is empty and has no columns.

We check for the size again later anyway, so we can just check it again now
before printing out values.