Handle the case of multiple stopped_moving transitions
with no intermediate `exited_geofence` transitions
This basically means that we can have places where the duration is
less than 5 minutes. Which means that setting the start time to the
place enter + 3 minutes can cause the start to occur after the first recorded
point.

Have asked the user for phone logs.
Workaround is to move ahead by half the duration or 3 secs, whichever is
smaller.

Without this, processing data from the 7th for this user in two batches
(before and after 18:33:11) throws IndexError, With this, there is no such
error.

On directly deploying this to production, the previously failing trip was
processed successfully.

Also added some additional debug statements to segmentation code
https://github.com/e-mission/e-mission-server/issues/436#issuecomment-253114786