Switch the stats to the timeseries, similar to the other data
Back when we stored the raw data directly as sections and trips, we used to
store the stats in a timeseries. Now that we are storing the raw data in a
timeseries, and are planning to use the stats for analysis into user behavior,
it would be more convenient to store the stats into the same timeseries. Then,
the timeseries is really the a full representation of the source of truth, we
can reuse similar code for analysis.

This change:
- defines a new wrapper class for the data
- supports it in the entry and the timeseries
- switches the server api tracking to the new timeseries
- adds a backwards compat layer that stores old-style client stats to the new timeseries, while waiting for the clients to get updated

In later changes:
- migration script to migrate the older style stats to the new timeseries
- new stats for the pipeline