Restore the geojson properties to a section
Before this, we converted the section's sensed_mode to a string to get
`ON_FOOT`, etc. However, when we merged the section with the previous
properties (to account for the times), we ended up with a dict, which returned
the value for the mode.

This restores the properties as a wrapper object and makes it possible to read
the sensed mode correctly again.