Populate the timestamp into the data from the metadata only if it doesn't exist
We now populate the date in the transition, so we can retain it instead of
populating it from the metadata. While doing that, refactor the code to support
filling in additional time information - local_dt, fmt_time for the data and use
it everywhere.

Commit that populates the date in the transition
https://github.com/e-mission/e-mission-data-collection/commit/898f2792939d417257616b725200957c5afb77bf