Ignore invalid points where the data.ts and the metadata.write_ts differ
In https://github.com/e-mission/e-mission-server/issues/457, we notice that we
can sometimes get invalid points in which the data.ts and the metadata.write_ts
differ by a lot. I suspect that this is due to manual manipulation of the
datetime. If this occurs only once or twice, it is not a big deal, but if it
happens for a period of time, we can actually go backwards in time during
segmentation, which results in all kinds of weird problems.

Also add unit test case and ground truth