Moves the giles archiver test out of te way
We have run out of space on server running the new and improved stack.
This means that the tests fail.
We have already stopped using this (all data is stored in the timeseries) and
moved out the stats tests. Let's move this out as well.

We can then shut down the server and re-provision something to run the open
parse server for push notifications.