Switch to a parse compatible push notification server
When using our own server, I was able to connect to it from another AWS server
and from my laptop, but wasn't able to connect from the phone. Not even from
Safari, so this is not some weird app permission issue. Need to play around
with it a bit more, but for now, let's use a parse-compatible hosted provider instead.

This has been tested by making changes directly in production and it works.