Add a new script that downloads data for a date range
Because this is a date range, we can no longer use localdate (which is a
filtering technique).
Instead we convert to a timestamp and use timestamps instead.
We also query using the timeseries, so we get both raw and analysed data.
We need to include multiple time queries to read location data using `data.ts`,
trip-like data using `data.start_ts` and place-like data using `data.enter_ts`.