Fix the implementation of `get_complete_ts`
The last time this was changed was on Oct 4, 2015, when we weren't tracking the
`last_processed_ts` and didn't have the cleaned infrastructure.
https://github.com/e-mission/e-mission-server/commit/6992da8c7ed7e809872deb1f3dafc31639c4f1c2

We added the last_processed timestamp on Oct 8th, 2015.
https://github.com/e-mission/e-mission-server/commit/ec004153d40a61e10097e01debabb76bf1cdf444

and the CLEANED_RESAMPLING state on Apr 22, 2016
https://github.com/e-mission/e-mission-server/commit/36b832629f3db7711bc03a7640db48a50edd4558

Looks like we just forgot to update this timestamp, but better late than never!