Save the timeline into a set of user files
For defensive reasons (to ensure that sending in a very broad query does not
crash the server/take too much time), I limit the number of results returned by
find_entries
https://github.com/e-mission/e-mission-server/blob/master/emission/storage/timeseries/builtin_timeseries.py#L213

That limit is fine for a single user, but when I pull data for multiple users
at a time, we hit the limit and so some data is not dumped.  This is also why I
didn't run into it when I tested against a smaller range earlier.

Fix it by querying for each user separately. Then we can also save the data for
each user into a separate file. This is much more memory-safe and performant
since we don't have to have all the data in memory at the same time, and
ensures that we don't run out of memory while loading a large dataset.

We then need to fix the load and purge scripts to work off multiple files as well.