Add support to reset habitica autocheck tasks to a particular point
Currently, this assumes that points were not given after that point.
If points were given, but incorrectly, we don't have a huge option on how to
fix it since there is no way to "set points".

TODO: investigate whether decrementing the task counter the appropriate number
of times give an idempotent result?