Add support for the token list authenticator
This is pretty simple - it has a hardcoded list of tokens, and it checks the
incoming token against the hardcoded list. If there is a match, it is
authenticated, if there is no match, a ValueError is generated.

I tried just returning None instead of generating a ValueError, but then it
just gets mapped to a user with email = None.