Move the keys.json file into a sample file
Similar to 48e5a60cea216e9c03d75b8beea41a177c47cc6c.
Allows us to add entries without causing merge conflicts
Note that we don't add fallback code to read from the sample file if the real
one does not exist. This is because the most auth (except for skip) will fail
if not configured properly, and we want to fail fast in that case