Initial version of the switch from ionic to firebase for the push notifications
This is essentially the same code as the one that is running on the server right now.
It was generated by a patch from the server.
The only change is to replace the hardcoded API key on the server with the
value read from the config file, to avoid leaking that information.